# Conflicts:
#	src/main/java/com/utn/space/venueaapi/model/Space.java
#	src/main/java/com/utn/space/venueaapi/repository/SpaceRepository.java

+++ src/main/java/com/utn/space/venueaapi/exceptions/ExceptionIdNotFound.java
@@ -1,7 +1,7 @@
 package com.utn.space.venueaapi.exceptions;
 
 public class ExceptionIdNotFound extends RuntimeException {
-    public ExceptionIdNotFound(String objet,Long id) {
+    public ExceptionIdNotFound(String objet,Integer id) {
         super("El id: " + id + " de " + objet + " no existe.");
     }
 }
